add UI for games and authors
