+++ games/choices.rb
@@ -1,0 +1,9 @@
+require_relative './create_game'
+
+def when_yes
+  implement_choice
+end
+
+def when_no
+  implement_choice
+end
