+++ store/authors.json
@@ -1,0 +1,10 @@
+[
+  {
+    "first_name": "Milen",
+    "last_name": "Alebel"
+  },
+  {
+    "first_name": "Christian",
+    "last_name": "Iradukunda"
+  }
+]
