+++ store/games.json
@@ -1,0 +1,16 @@
+[
+  {
+    "name_of_game": "Cookie Jam",
+    "multiplayer": "Yes",
+    "last_played_at": "1",
+    "first_name": "Milen",
+    "last_name": "Alebel"
+  },
+  {
+    "name_of_game": "COD",
+    "multiplayer": "Yes",
+    "last_played_at": "6",
+    "first_name": "Christian",
+    "last_name": "Iradukunda"
+  }
+]
